Use envelope rcptTo instead of content to header for db collection
Emails that don't have a to header aren't handled. For instance emails that only have bcc
Use the envelope rcpto values instead of the to header for the db collection which will always exist
Add a dummy to value if one doesn't exist. Several places assume there is one
Remove all unneeded mongo filters on the to value as each address already has its own collection
Store email for each rcpto if there are multiples